Add more push param tests
