+++ tests/Messaging/Adapter/Push/APNSTest.php
@@ -3,36 +3,29 @@
 namespace Utopia\Tests\Adapter\Push;
 
 use Utopia\Messaging\Adapter\Push\APNS as APNSAdapter;
-use Utopia\Messaging\Messages\Push;
-use Utopia\Tests\Adapter\Base;
 
 class APNSTest extends Base
 {
-    public function testSend(): void
+    protected function setUp(): void
     {
+        parent::setUp();
+
         $authKey = \getenv('APNS_AUTHKEY_8KVVCLA3HL');
         $authKeyId = \getenv('APNS_AUTH_ID');
         $teamId = \getenv('APNS_TEAM_ID');
         $bundleId = \getenv('APNS_BUNDLE_ID');
 
-        $adapter = new APNSAdapter($authKey, $authKeyId, $teamId, $bundleId, true);
-
-        $message = new Push(
-            to: [\getenv('APNS_TO')],
-            title: 'Test title',
-            body: 'Test body',
-            data: null,
-            action: null,
-            sound: 'default',
-            icon: null,
-            color: null,
-            tag: null,
-            badge: 1,
-            contentAvailable: true
+        $this->adapter = new APNSAdapter(
+            $authKey,
+            $authKeyId,
+            $teamId,
+            $bundleId,
+            sandbox: true
         );
+    }
 
-        $response = $adapter->send($message);
-
-        $this->assertResponse($response);
+    protected function getTo(): array
+    {
+        return [\getenv('APNS_TO')];
     }
 }
